Feat: Fully integrate cyclic finding into normal Finder
