--- tests/org.dataflowanalysis.analysis.tests/src/org/dataflowanalysis/analysis/tests/sequencefinder/CyclicOnlineShopTest.java
@@ -1,13 +1,0 @@
-package org.dataflowanalysis.analysis.tests.sequencefinder;
-
-import org.dataflowanalysis.analysis.core.TransposeFlowGraphFinder;
-import org.dataflowanalysis.analysis.dfd.core.DFDCyclicTransposeFlowGraphFinder;
-
-public class CyclicOnlineShopTest extends BaseOnlineShopTest{
-
-    @Override
-    protected Class<? extends TransposeFlowGraphFinder> getTransposeFlowGraphFinder() {
-        return DFDCyclicTransposeFlowGraphFinder.class;
-    }
-
-}
